Replace ParserNonS with ParserNonS2
