+++ hs-src/Language/Egison/Core.hs
@@ -66,7 +66,6 @@ import qualified Data.Text                   as T
 
 import           Language.Egison.Parser      as Parser
 import           Language.Egison.ParserNonS  as ParserNonS
-import           Language.Egison.ParserNonS2 as ParserNonS2
 import           Language.Egison.Types
 
 --
@@ -84,7 +83,6 @@ collectDefs opts (expr:exprs) bindings rest =
       if optNoIO opts
          then throwError $ Default "No IO support"
          else do exprs' <- if
-                   | optUseNonS2 opts -> ParserNonS2.loadFile file
                    | optSExpr opts    -> Parser.loadFile file
                    | otherwise        -> ParserNonS.loadFile file
                  collectDefs opts (exprs' ++ exprs) bindings rest
@@ -92,7 +90,6 @@ collectDefs opts (expr:exprs) bindings rest =
       if optNoIO opts
          then throwError $ Default "No IO support"
          else do exprs' <- if
-                   | optUseNonS2 opts -> ParserNonS2.loadLibraryFile file
                    | optSExpr opts    -> Parser.loadLibraryFile file
                    | otherwise        -> ParserNonS.loadLibraryFile file
                  collectDefs opts (exprs' ++ exprs) bindings rest
