+++ hs-src/Language/Egison/Util.hs
@@ -19,7 +19,6 @@ import           Text.Regex.TDFA                  ((=~))
 
 import           Language.Egison.Parser           as Parser
 import           Language.Egison.ParserNonS       as ParserNonS
-import           Language.Egison.ParserNonS2      as ParserNonS2
 import           Language.Egison.Types
 
 -- |Get Egison expression from the prompt. We can handle multiline input.
@@ -41,7 +40,7 @@ getEgisonExpr' opts prev = do
       history <- getHistory
       putHistory $ addHistoryUnlessConsecutiveDupe line history
       let input = prev ++ line
-      let parsedExpr = (if optUseNonS2 opts then ParserNonS2.parseTopExpr else if optSExpr opts then Parser.parseTopExpr else ParserNonS.parseTopExpr) input
+      let parsedExpr = (if optSExpr opts then Parser.parseTopExpr else ParserNonS.parseTopExpr) input
       case parsedExpr of
         Left err | show err =~ "unexpected end of input" ->
           getEgisonExpr' opts $ input ++ "\n"
